refactor: complete P1 architecture (match, extract, form, god-files)
Shrink remaining God-Files, consolidate match under lib/match, add
document-text facade with extract types in lib, and decouple reference
form fields/types from the dashboard adapter layer.
