+++ app/dashboard/deals/components/match-result-card.tsx
@@ -13,7 +13,7 @@ import { cn } from '@/lib/utils'
 import { MatchScoreCircle } from '@/components/match/match-score-circle'
 import { getMatchStrength } from '@/lib/match/match-strength'
 import { formatMatchReferenceMetaLine } from '@/lib/match/match-reference-meta'
-import type { MatchReferenceHit } from '@/app/dashboard/actions'
+import type { MatchReferenceHit } from '@/lib/match/match-types'
 import { createSharedPortfolio } from '@/app/dashboard/actions'
 import { addReferenceToDealWithScore, removeReferenceFromDeal } from '../actions'
 import { PdfExportDialog } from '@/app/dashboard/references/[id]/pdf-export-dialog'
