+++ app/dashboard/references/new/reference-form.tsx
@@ -1,9 +1,27 @@
 'use client'
 
+import type { ComponentProps } from 'react'
+import {
+  ReferenceForm as ReferenceFormBase,
+} from '@/lib/references/reference-form/reference-form'
+import { searchCompanySuggestions } from './actions'
+
 export type {
   ContactPerson,
   ExternalContactDisplay,
   ReferenceFormInitialData,
 } from '@/lib/references/reference-form/reference-form-types'
 
-export { ReferenceForm } from '@/lib/references/reference-form/reference-form'
+/** Dashboard-Adapter: injiziert Org-Suche; Lib bleibt von Actions entkoppelt. */
+export function ReferenceForm(
+  props: Omit<ComponentProps<typeof ReferenceFormBase>, 'searchCompanies'> & {
+    searchCompanies?: ComponentProps<typeof ReferenceFormBase>['searchCompanies']
+  }
+) {
+  return (
+    <ReferenceFormBase
+      {...props}
+      searchCompanies={props.searchCompanies ?? searchCompanySuggestions}
+    />
+  )
+}
