+++ app/dashboard/smart-match/smart-match-shell-helpers.ts
@@ -1,8 +1,5 @@
 import { MASTER_INDUSTRIES, formatIndustryDisplay, resolveIndustryId } from '@/lib/constants/industries'
-import {
-  type MatchReferenceHit,
-  type MatchReferenceFilters,
-} from '@/app/dashboard/actions'
+import type { MatchReferenceHit, MatchReferenceFilters } from '@/lib/match/match-types'
 import type { DealRow } from '@/app/dashboard/deals/types'
 import { parseStoredVolumeEur } from '@/lib/match/parse-smart-match-query'
 import type { ParsedSmartMatchConstraints } from '@/lib/match/parse-smart-match-query'
