+++ lib/command-center/search-references-semantic.ts
@@ -10,8 +10,8 @@ import {
 } from '@/lib/command-center/homepage-semantic-query'
 import type { HomepageSemanticReferenceHit } from '@/lib/command-center/homepage-semantic-types'
 import { embedTextWithOpenAICached } from '@/lib/embeddings/cached-embed-query'
-import { rpcMatchReferences } from '@/lib/match-references-rpc'
-import { snippetFromSummary } from '@/lib/match-reference-snippet'
+import { rpcMatchReferences } from '@/lib/match/match-references-rpc'
+import { snippetFromSummary } from '@/lib/match/match-reference-snippet'
 import { fetchCompanyFieldsForReferenceIds } from '@/lib/references/enrich-match-hits-company'
 import { logReferenceMatched } from '@/lib/events/log-reference-matched'
 import { log } from '@/lib/observability/logger'
