+++ lib/references/library/bulk-import.ts
@@ -6,7 +6,7 @@ import { ROUTES } from '@/lib/routes'
 import { revalidateOrgReferences } from '@/lib/cache/revalidate-org'
 import { BULK_IMPORT_MAX_FILES } from '@/lib/references/bulk-import-limits'
 import { resolveOrCreateCompanyForImport } from '@/lib/accounts/resolve-company-for-import'
-import { extractPlainTextFromBuffer } from '@/lib/document-extraction'
+import { extractPlainTextFromBuffer } from '@/lib/document-text'
 import { parseReferenceHeuristicsFromText } from '@/lib/references/heuristic-reference-extract'
 import { normalizeNarrativeText } from '@/lib/references/narrative-normalize'
 import { mapBrandfetchIndustriesArrayToGermanCategory } from '@/lib/brandfetch/map-brandfetch-industry-to-de'
